Make test suites for integration tests to minimize amount of db connections
